+++ app/src/main/java/app/knotwork/android/presentation/ui/automation/ExternalAutomationJournalScreen.kt
@@ -2,13 +2,11 @@ package app.knotwork.android.presentation.ui.automation
 
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.material3.SnackbarHost
 import androidx.compose.material3.SnackbarHostState
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.produceState
 import androidx.compose.runtime.remember
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalConfiguration
 import androidx.compose.ui.platform.LocalContext
@@ -30,6 +28,7 @@ import app.knotwork.android.domain.usecases.JournalTimestamp
 import app.knotwork.android.domain.usecases.automation.CleanupExternalAutomationJournalUseCase
 import app.knotwork.android.presentation.ui.common.rememberJournalExportHandlers
 import app.knotwork.android.presentation.ui.common.writePlainClipboardText
+import app.knotwork.design.components.misc.KnotworkSnackbarHost
 import app.knotwork.design.screens.automation.ExternalAutomationJournalCallbacks
 import app.knotwork.design.screens.automation.ExternalAutomationJournalContent
 import app.knotwork.design.screens.automation.ExternalAutomationJournalStrings
@@ -106,8 +105,8 @@ fun ExternalAutomationJournalScreen(
                 onShareJournal = journalExport.onShare,
                 onSaveJournal = journalExport.onSave,
             ),
+            snackbarHost = { KnotworkSnackbarHost(hostState = snackbarHostState) },
         )
-        SnackbarHost(hostState = snackbarHostState, modifier = Modifier.align(Alignment.BottomCenter))
     }
 }